VMM/IEM: ARM build fix. bugref:10371
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@102734 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/include/IEMN8veRecompilerEmit.h
@@ -3377,7 +3377,7 @@ iemNativeEmitJccToLabelEx(PIEMRECOMPILERSTATE pReNative, PIEMNATIVEINSTR pCodeBu
     }
     else
     {
-        Assert(offLabel - off <= -0x3ffff);
+        Assert((int64_t)(offLabel - off) <= -0x3ffff);
         pCodeBuf[off++] = Armv8A64MkInstrBCond(enmCond, offLabel - off);
     }
